requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
alpaca-trade-api==3.3.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
websocket-client==1.8.0
